+++ src/server/modes/base/responses/broadcast/chat-server-public.ts
@@ -1,4 +1,5 @@
-import { BROADCAST_CHAT_PUBLIC, BROADCAST_CHAT_SERVER_PUBLIC } from '@/events';
+import { ServerPackets, SERVER_PACKETS } from '@airbattle/protocol';
+import { BROADCAST_CHAT_SERVER_PUBLIC, CONNECTIONS_SEND_PACKET } from '@/events';
 import { System } from '@/server/system';
 
 export default class ChatServerPublicBroadcast extends System {
@@ -16,13 +17,18 @@ export default class ChatServerPublicBroadcast extends System {
    * @param text
    */
   onChatServerPublic(text: string): void {
+    const recipients = [...this.storage.humanConnectionIdList];
     let offset = 0;
 
     while (offset < text.length) {
       this.emit(
-        BROADCAST_CHAT_PUBLIC,
-        this.storage.serverPlayerId,
-        text.substring(offset, offset + 255)
+        CONNECTIONS_SEND_PACKET,
+        {
+          c: SERVER_PACKETS.CHAT_PUBLIC,
+          id: this.storage.serverPlayerId,
+          text: text.substring(offset, offset + 255),
+        } as ServerPackets.ChatPublic,
+        recipients
       );
 
       offset += 255;