+++ CHANGES.md
@@ -5,6 +5,10 @@ Features:
 - Command /spectators. Shows how many players are spectating you.
 - Customise server bot by passing environment variables SERVER_BOT_NAME and SERVER_BOT_FLAG.
 
+Improvements:
+
+- The server bot is completely "removed" for other bots.
+
 ## 4.6.0 (September 19, 2019)
 
 Improvements: